Updated events to cater for transaction hash
