+++ package.json
@@ -21,10 +21,10 @@
   },
   "homepage": "https://github.com/Glitter-Finance/glitter-bridge-connections#readme",
   "dependencies": {
+    "@circle-fin/circle-sdk": "^1.3.0",
     "@perawallet/connect": "^1.0.7",
     "@project-serum/sol-wallet-adapter": "^0.2.6",
     "@randlabs/myalgo-connect": "^1.4.2",
-    "@circle-fin/circle-sdk": "^1.3.0",
     "@solana/spl-token": "^0.3.6",
     "@solana/web3.js": "^1.65.0",
     "algosdk": "^1.24.1",
@@ -33,8 +33,8 @@
     "dotenv": "^16.0.3",
     "ed25519-hd-key": "^1.3.0",
     "ethers": "^5.7.2",
-    "glitter-evm-contracts": "^1.0.31",
     "eventemitter3": "^5.0.0",
+    "glitter-evm-contracts": "^1.0.32",
     "minimist": "latest",
     "tweetnacl": "^1.0.3",
     "winston": "latest"
